[Cli Ophthalmology] Update vision test field types change from integer to decimal in sphere and axis

+++ services/Ophthalmology/clinic_app/models.py
@@ -5,11 +5,11 @@ from django.contrib.postgres.fields import ArrayField
 class VisionTest(models.Model):
     id = models.BigAutoField(primary_key=True)
     appointment_id = models.IntegerField()
-    lt_sph = models.IntegerField()
-    lt_cyl = models.IntegerField()
+    lt_sph = models.DecimalField()
+    lt_cyl = models.DecimalField()
     lt_axis = models.IntegerField()
-    rt_sph = models.IntegerField()
-    rt_cyl = models.IntegerField()
+    rt_sph = models.DecimalField()
+    rt_cyl = models.DecimalField()
     rt_axis = models.IntegerField()
     def __str__(self):
         return f'id {self.id} appt id {self.appointment_id}'
